+++ Changelog.md
@@ -57,6 +57,7 @@ v2.37.0 (not published yet)
 - NVEncC: Add "--vpp-nvvfx-denoise" parameter ([#1240](/../../issues/1240))
 - NVEncC: Fix "--vpp-resize" parameter setting for "nvvfx-superres"
 - VCEEncC: Fix problem with missing space before parameter call
+- VCEEncC: Remove Dolby Vision support
 - Update tools
     - ...
 - Update AviSynth+ plugins